+++ includes/data-events/connectors/class-mailchimp.php
@@ -63,6 +63,20 @@ class Mailchimp {
 	}
 
 	/**
+	 * Get default reader MailChimp status.
+	 *
+	 * @return string MailChimp status slug, 'transactional' or 'subscriber'. (Default: 'transactional').
+	 */
+	private static function get_default_reader_status() {
+		$allowed_statuses = [
+			'transactional',
+			'subscribed',
+		];
+		$default_status = Reader_Activation::get_setting( 'mailchimp_reader_default_status' );
+		return in_array( $default_status, $allowed_statuses, true ) ? $default_status : 'transactional';
+	}
+
+	/**
 	 * Get merge field type.
 	 *
 	 * @param mixed $value Value to check.
@@ -156,6 +170,14 @@ class Mailchimp {
 			);
 			// Skip field if it failed to create.
 			if ( is_wp_error( $created_field ) ) {
+				Logger::log(
+					sprintf(
+					// Translators: %1$s is the merge field key, %2$s is the error message.
+						__( 'Failed to create merge field %1$s. Error response: %2$s', 'newspack-plugin' ),
+						$field_name,
+						$created_field->get_error_message() ?? __( 'The connected ESP could not create this merge field.', 'newspack-plugin' )
+					)
+				);
 				continue;
 			}
 			Logger::log(
@@ -186,7 +208,7 @@ class Mailchimp {
 		$hash    = md5( strtolower( $contact['email'] ) );
 		$payload = [
 			'email_address' => $contact['email'],
-			'status_if_new' => 'transactional',
+			'status_if_new' => self::get_default_reader_status(),
 		];
 
 		// Normalize contact metadata.
@@ -291,22 +313,11 @@ class Mailchimp {
 	 * @param int   $client_id ID of the client that triggered the event.
 	 */
 	public static function subscription_updated( $timestamp, $data, $client_id ) {
-		if ( empty( $data['status_before'] ) || empty( $data['status_after'] ) || empty( $data['user_id'] ) ) {
+		if ( empty( $data['status_before'] ) || empty( $data['status_after'] ) || empty( $data['subscription_id'] ) ) {
 			return;
 		}
 
-		/*
-		 * If the subscription is being activated after a successful first or renewal payment,
-		 * the contact will be synced when that order is completed, so no need to sync again.
-		 */
-		if (
-			( 'pending' === $data['status_before'] || 'on-hold' === $data['status_before'] ) &&
-			'active' === $data['status_after'] ) {
-			return;
-		}
-
-		$customer = new \WC_Customer( $data['user_id'] );
-		$contact  = WooCommerce_Connection::get_contact_from_customer( $customer );
+		$contact = WooCommerce_Connection::get_contact_from_order( $data['subscription_id'] );
 
 		if ( ! $contact ) {
 			return;