+++ includes/data-events/connectors/ga4/class-ga4.php
@@ -270,6 +270,9 @@ class GA4 {
 			$params = array_merge( $params, self::get_sanitized_popup_params( $data['metadata']['newspack_popup_id'] ) );
 		}
 		if ( ! empty( $data['metadata']['referer'] ) ) {
+			$params['referrer'] = substr( $data['metadata']['referer'], 0, 100 );
+
+			// Retain both instances of referrer spelling to ensure publisher reports are not broken.
 			$params['referer'] = substr( $data['metadata']['referer'], 0, 100 );
 		}
 		return $params;
@@ -288,6 +291,8 @@ class GA4 {
 		$params['currency']        = $data['currency'];
 		$params['recurrence']      = $data['recurrence'];
 		$params['platform']        = $data['platform'];
+		$params['referrer']        = $data['referer'] ?? '';
+		// Retain both instances of referrer spelling to ensure publisher reports are not broken.
 		$params['referer']         = $data['referer'] ?? '';
 		$params['popup_id']        = $data['popup_id'] ?? '';
 		$params['is_renewal']      = $data['is_renewal'] ? 'yes' : 'no';
@@ -354,7 +359,10 @@ class GA4 {
 			$params = array_merge( $params, self::get_sanitized_popup_params( $metadata['newspack_popup_id'] ) );
 		}
 		$params['newsletters_subscription_method'] = $metadata['newsletters_subscription_method'] ?? '';
-		$params['referer']                         = $metadata['current_page_url'] ?? '';
+		$params['referrer']                        = $metadata['current_page_url'] ?? '';
+
+		// Retain both instances of referrer spelling to ensure publisher reports are not broken.
+		$params['referer'] = $metadata['current_page_url'] ?? '';
 
 		// In case the subscription happened as part of the registration process, we should also have the registration method.
 		$params['registration_method'] = $metadata['registration_method'] ?? '';
@@ -397,6 +405,8 @@ class GA4 {
 		$params['gate_post_id'] = $data['gate_post_id'] ?? '';
 		$params['action']       = $data['action'] ?? '';
 		$params['action_type']  = $data['action_type'] ?? '';
+		$params['referrer']     = $data['referer'] ?? '';
+		// Retain both instances of referrer spelling to ensure publisher reports are not broken.
 		$params['referer']      = $data['referer'] ?? '';
 		$params['order_id']     = $data['order_id'] ?? '';
 		$params['product_id']   = $data['product_id'] ?? '';