+++ includes/plugins/class-newspack-newsletters.php
@@ -17,6 +17,12 @@ class Newspack_Newsletters {
 	const METADATA_PREFIX        = 'NP_';
 	const METADATA_PREFIX_OPTION = '_newspack_metadata_prefix';
 
+	/**
+	 * The option name for choosing which metadata fields to sync.
+	 *
+	 * @var string
+	 */
+	const METADATA_FIELDS_OPTION = '_newspack_metadata_fields';
 
 	/**
 	 * Metadata keys map for Reader Activation.
@@ -54,6 +60,13 @@ class Newspack_Newsletters {
 	 * Initialize hooks and filters.
 	 */
 	public static function init() {
+		/**
+		 * Filters the list of key/value pairs for metadata fields to be synced to the connected ESP.
+		 *
+		 * @param array $metadata_keys The list of key/value pairs for metadata fields to be synced to the connected ESP.
+		 */
+		self::$metadata_keys = \apply_filters( 'newspack_ras_metadata_keys', self::$metadata_keys );
+
 		\add_filter( 'newspack_newsletters_contact_data', [ __CLASS__, 'normalize_contact_data' ] );
 
 		if ( self::should_sync_ras_metadata() ) {
@@ -108,6 +121,73 @@ class Newspack_Newsletters {
 	}
 
 	/**
+	 * Get the list of possible fields to be synced to the connected ESP.
+	 *
+	 * @return string[] List of fields.
+	 */
+	public static function get_default_metadata_fields() {
+		return array_values( array_unique( array_values( self::$metadata_keys ) ) );
+	}
+
+	/**
+	 * Get the list of fields to be synced to the connected ESP.
+	 *
+	 * @return string[] List of fields to be synced.
+	 */
+	public static function get_metadata_fields() {
+		return array_values( \get_option( self::METADATA_FIELDS_OPTION, self::get_default_metadata_fields() ) );
+	}
+
+	/**
+	 * Update the list of fields to be synced to the connected ESP.
+	 *
+	 * @param array $fields List of fields to sync.
+	 *
+	 * @return boolean True if updated, false otherwise.
+	 */
+	public static function update_metadata_fields( $fields ) {
+		// Only allow fields that are in the metadata keys map.
+		$fields = array_intersect( self::get_default_metadata_fields(), $fields );
+		return \update_option( self::METADATA_FIELDS_OPTION, array_values( $fields ) );
+	}
+
+	/**
+	 * Get the "raw" unprefixed metadata keys. Only return fields selected to sync to the ESP.
+	 *
+	 * @return string[] List of raw metadata keys.
+	 */
+	public static function get_raw_metadata_keys() {
+		$fields_to_sync = self::get_metadata_fields();
+		$raw_keys       = [];
+
+		foreach ( self::$metadata_keys as $raw_key => $field_name ) {
+			if ( in_array( $field_name, $fields_to_sync, true ) ) {
+				$raw_keys[] = $raw_key;
+			}
+		}
+
+		return array_unique( $raw_keys );
+	}
+
+	/**
+	 * Get the "prefixed" metadata keys. Only return fields selected to sync to the ESP.
+	 *
+	 * @return string[] List of prefixed metadata keys.
+	 */
+	public static function get_prefixed_metadata_keys() {
+		$fields_to_sync = self::get_metadata_fields();
+		$prefixed_keys  = [];
+
+		foreach ( self::$metadata_keys as $raw_key => $field_name ) {
+			if ( in_array( $field_name, $fields_to_sync, true ) ) {
+				$prefixed_keys[] = self::get_metadata_key( $raw_key );
+			}
+		}
+
+		return array_unique( $prefixed_keys );
+	}
+
+	/**
 	 * Given a field name, prepend it with the metadata field prefix.
 	 *
 	 * @param string $key Metadata field to fetch.
@@ -143,13 +223,8 @@ class Newspack_Newsletters {
 		// If syncing for RAS, ensure that metadata keys are normalized with the correct RAS metadata keys.
 		if ( isset( $contact['metadata'] ) ) {
 			$normalized_metadata = [];
-			$raw_keys            = array_values( array_flip( self::$metadata_keys ) );
-			$prefixed_keys       = array_map(
-				function( $key ) {
-					return self::get_metadata_key( $key );
-				},
-				$raw_keys
-			);
+			$raw_keys            = self::get_raw_metadata_keys();
+			$prefixed_keys       = self::get_prefixed_metadata_keys();
 
 			// Capture UTM params and signup/payment page URLs as meta for registration or payment.
 			if (
@@ -192,7 +267,11 @@ class Newspack_Newsletters {
 						$normalized_metadata[ self::get_metadata_key( $meta_key ) ] = $meta_value; // If passed a raw key, map it to the prefixed key.
 					} elseif (
 						in_array( $meta_key, $prefixed_keys, true ) ||
-						( false !== strpos( $meta_key, self::get_metadata_key( 'signup_page_utm' ) ) || false !== strpos( $meta_key, self::get_metadata_key( 'payment_page_utm' ) ) ) // UTM meta keys can have arbitrary suffixes.
+						(
+							// UTM meta keys can have arbitrary suffixes.
+							( in_array( self::get_metadata_key( 'signup_page_utm' ), $prefixed_keys, true ) && false !== strpos( $meta_key, self::get_metadata_key( 'signup_page_utm' ) ) ) ||
+							( in_array( self::get_metadata_key( 'payment_page_utm' ), $prefixed_keys, true ) && false !== strpos( $meta_key, self::get_metadata_key( 'payment_page_utm' ) ) )
+						)
 					) {
 						$normalized_metadata[ $meta_key ] = $meta_value;
 					}
@@ -203,11 +282,16 @@ class Newspack_Newsletters {
 					}
 				}
 			}
+
+			// Ensure status is passed, if given.
+			if ( isset( $contact['metadata']['status'] ) ) {
+				$normalized_metadata['status'] = $contact['metadata']['status'];
+			}
 			$contact['metadata'] = $normalized_metadata;
 		}
 
 		// Parse full name into first + last for MC, which stores these as separate merge fields.
-		if ( method_exists( 'Newspack_Newsletters', 'service_provider' ) && 'mailchimp' === \Newspack_Newsletters::service_provider() ) {
+		if ( 'mailchimp' === \get_option( 'newspack_newsletters_service_provider', false ) ) {
 			if ( isset( $contact['name'] ) ) {
 				if ( ! isset( $contact['metadata'] ) ) {
 					$contact['metadata'] = [];
@@ -223,7 +307,12 @@ class Newspack_Newsletters {
 		Logger::log( 'Normalizing contact data for reader ESP sync:' );
 		Logger::log( $contact );
 
-		return $contact;
+		/**
+		 * Filters the normalized contact data before syncing to the ESP.
+		 *
+		 * @param array $contact Contact data.
+		 */
+		return apply_filters( 'newspack_esp_sync_normalize_contact', $contact );
 	}
 
 	/**